+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fixing bug creating non productive forest when expanding verified classes
+- Handled properly when user asks for an random object of a size the class does
+  not have
 
 ## [4.0.0] - 2021-06-14
 ### Added